requirements.txt
# Conecta Beauty - Dependências

streamlit>=1.28.0
pandas>=2.0.0
pymongo>=4.5.0
plotly>=5.17.0
requests>=2.31.0
agno>=0.1.0
python-dotenv>=1.0.0
